-changed the login function (made it better)
